requirements.txt
mysql-connector-python>=8.2.0
pandas>=2.1.4
numpy>=1.24.4
python-dotenv>=1.1.1